ruqya-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
websockets==12.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.34.34
resend==0.7.0
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.23.4
python-slugify==8.0.1
python-dotenv==1.0.0